+++ app/src/androidTest/java/fi/thl/koronahaavi/FakeExposureRepository.kt
@@ -1,6 +1,7 @@
 package fi.thl.koronahaavi
 
 import fi.thl.koronahaavi.data.Exposure
+import fi.thl.koronahaavi.data.ExposureNotification
 import fi.thl.koronahaavi.data.ExposureRepository
 import fi.thl.koronahaavi.data.KeyGroupToken
 import kotlinx.coroutines.flow.Flow
@@ -8,15 +9,12 @@ import kotlinx.coroutines.flow.flowOf
 import javax.inject.Inject
 
 class FakeExposureRepository : ExposureRepository {
-    override suspend fun getAllKeyGroupTokens(): List<KeyGroupToken>
-            = listOf()
+    override suspend fun getAllKeyGroupTokens(): List<KeyGroupToken> = listOf()
 
     override suspend fun saveKeyGroupToken(token: KeyGroupToken) {
     }
 
-    override fun flowHandledKeyGroupTokens(): Flow<List<KeyGroupToken>> {
-        return flowOf(listOf())
-    }
+    override fun flowHandledKeyGroupTokens(): Flow<List<KeyGroupToken>> = flowOf(listOf())
 
     override suspend fun deleteKeyGroupToken(token: KeyGroupToken) {
     }
@@ -24,13 +22,8 @@ class FakeExposureRepository : ExposureRepository {
     override suspend fun saveExposure(exposure: Exposure) {
     }
 
-    override suspend fun getExposure(id: Long): Exposure? {
-        return null
-    }
-
-    override suspend fun getAllExposures(): List<Exposure> {
-        return listOf()
-    }
+    override suspend fun getExposure(id: Long): Exposure? = null
+    override suspend fun getAllExposures(): List<Exposure> = listOf()
 
     override suspend fun deleteAllExposures() {
     }
@@ -38,11 +31,9 @@ class FakeExposureRepository : ExposureRepository {
     override suspend fun deleteExposure(id: Long) {
     }
 
-    override fun flowAllExposures(): Flow<List<Exposure>> {
-        return flowOf(listOf())
-     }
-
-    override fun flowHasExposures(): Flow<Boolean> {
-        return flowOf(false)
+    override suspend fun deleteExpiredExposuresAndTokens() {
     }
+
+    override fun flowAllExposures(): Flow<List<Exposure>> = flowOf(listOf())
+    override fun flowExposureNotifications(): Flow<List<ExposureNotification>> = flowOf(listOf())
 }