+++ app/src/main/java/fi/thl/koronahaavi/common/ENEnablerFragment.kt
@@ -10,7 +10,6 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.lifecycleScope
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.ConnectionResult.SERVICE_INVALID
-import com.google.android.gms.common.GoogleApiAvailability
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import dagger.hilt.android.AndroidEntryPoint
 import fi.thl.koronahaavi.R
@@ -18,6 +17,7 @@ import fi.thl.koronahaavi.common.RequestResolutionViewModel.Companion.REQUEST_CO
 import fi.thl.koronahaavi.settings.ENApiError
 import fi.thl.koronahaavi.settings.EnableENError
 import fi.thl.koronahaavi.settings.EnableSystemViewModel
+import fi.thl.koronahaavi.settings.UserEnableStep
 import kotlinx.coroutines.launch
 import timber.log.Timber
 
@@ -38,8 +38,14 @@ open class ENEnablerFragment : Fragment() {
 
         statusViewModel.enableErrorEvent().observe(viewLifecycleOwner, Observer {
             it.getContentIfNotHandled()?.let { reason ->
-                context?.showEnableFailureReasonDialog(reason)
-                onEnableCanceled()
+                if (reason is EnableENError.UserCanceled && reason.step is UserEnableStep.UserConsent) {
+                    // huawei contact shield reports cancel here instead of activity result through request resolution
+                    onUserRejectedEnable()
+                }
+                else {
+                    context?.showEnableFailureReasonDialog(reason)
+                    onEnableCanceled()
+                }
             }
         })
 
@@ -90,19 +96,19 @@ open class ENEnablerFragment : Fragment() {
      * enabling EN API.
      */
     protected fun startEnablingSystem() {
-        val gaa = GoogleApiAvailability.getInstance()
-        val result = gaa.isGooglePlayServicesAvailable(requireContext(), MIN_GOOGLE_PLAY_VERSION)
+        val resolver = statusViewModel.getSystemAvailability()
+        val result = resolver.isSystemAvailable(requireContext())
 
         if (result == ConnectionResult.SUCCESS) {
-            Timber.v("Play services up-to-date")
+            Timber.v("Exposure notification system services available")
             enableSystem()
-
-        } else {
-            Timber.i("Play services needs updating ($result)")
+        }
+        else {
+            Timber.i("Exposure notification system needs updating ($result)")
             // The shown dialog displays something informative and the action button can for example
             // take the user to Play store to update Play services, or open device's Play services
             // app settings to enable it.
-            val shown = gaa.showErrorDialogFragment(activity, result, REQUEST_CODE_PLAY_SERVICES_ERROR_DIALOG) {
+            val shown = resolver.showErrorDialogFragment(requireActivity(), result, REQUEST_CODE_PLAY_SERVICES_ERROR_DIALOG) {
                 // callback if user backs out of or cancels the dialog
                 onEnableCanceled()
             }
@@ -110,9 +116,9 @@ open class ENEnablerFragment : Fragment() {
             if (!shown) {
                 enableSystem()
             } else {
-                Timber.d("Play services error dialog shown, result=$result, resolvable=${gaa.isUserResolvableError(result)}")
+                Timber.d("Play services error dialog shown, result=$result, resolvable=${resolver.isUserResolvableError(result)}")
 
-                if (result == SERVICE_INVALID || !gaa.isUserResolvableError(result)) {
+                if (result == SERVICE_INVALID || !resolver.isUserResolvableError(result)) {
                     // these unresolvable errors show a message dialog but will not post result
                     // codes to activity, so we need to invoke cancel callback to clear in-progress state
                     onEnableCanceled()
@@ -121,10 +127,6 @@ open class ENEnablerFragment : Fragment() {
                 // -> play dialog result is posted to playServicesResolvedEvent
             }
         }
-    }
-
-    companion object {
-        const val MIN_GOOGLE_PLAY_VERSION = 201813000   // v20.18.13
     }
 }
 
@@ -154,6 +156,18 @@ fun Context.showEnableFailureReasonDialog(error: EnableENError) {
                 else
                     getString(R.string.enable_err_api_not_supported, apiError)
             )
+        }
+
+        is EnableENError.UserCanceled -> {
+            // Huawei only
+            when (error.step) {
+                is UserEnableStep.LocationPermission -> {
+                    builder.setMessage(getString(R.string.enable_err_hms_location_permission))
+                }
+                else -> {
+                    return // should not get here, but handle anyway by skipping dialog
+                }
+            }
         }
 
         is EnableENError.Failed -> {