+++ app/src/main/java/fi/thl/koronahaavi/data/ExposureNotification.kt
@@ -1,0 +1,8 @@
+package fi.thl.koronahaavi.data
+
+import java.time.ZonedDateTime
+
+data class ExposureNotification(
+    val createdDate: ZonedDateTime,
+    val exposures: List<Exposure>
+)