+++ app/src/main/java/fi/thl/koronahaavi/service/ExposureNotificationService.kt
@@ -2,33 +2,33 @@
 
 package fi.thl.koronahaavi.service
 
+import android.app.Activity
+import android.content.Context
+import android.content.DialogInterface
 import com.google.android.gms.common.api.Status
-import com.google.android.gms.nearby.exposurenotification.ExposureInformation
 import com.google.android.gms.nearby.exposurenotification.ExposureSummary
 import com.google.android.gms.nearby.exposurenotification.TemporaryExposureKey
 import fi.thl.koronahaavi.data.Exposure
 import kotlinx.coroutines.flow.StateFlow
-import timber.log.Timber
 import java.io.File
-import java.time.Instant
-import java.time.ZoneOffset
-import java.time.ZonedDateTime
 
 interface ExposureNotificationService {
     suspend fun enable(): ResolvableResult<Unit>
     suspend fun disable(): ResolvableResult<Unit>
     suspend fun isEnabled(): Boolean
     suspend fun getExposureSummary(token: String): ExposureSummary
-    suspend fun getExposureDetails(token: String): List<ExposureInformation>
+    suspend fun getExposureDetails(token: String): List<Exposure>
     suspend fun provideDiagnosisKeyFiles(token: String, files: List<File>, config: ExposureConfigurationData): ResolvableResult<Unit>
     suspend fun getTemporaryExposureKeys(): ResolvableResult<List<TemporaryExposureKey>>
     fun deviceSupportsLocationlessScanning(): Boolean
+    fun getAvailabilityResolver(): AvailabilityResolver
 
     fun isEnabledFlow() : StateFlow<Boolean?>
     suspend fun refreshEnabledFlow()
 
     // represents different EN API method results: success, needs user resolution or failed
     sealed class ResolvableResult<T> {
+        data class HmsCanceled<T>(val step: EnableStep) : ResolvableResult<T>()
         data class ResolutionRequired<T>(val status: Status) : ResolvableResult<T>()
         data class Success<T>(val data: T) : ResolvableResult<T>()
         data class MissingCapability<T>(val error: String?) : ResolvableResult<T>()
@@ -44,14 +44,17 @@ interface ExposureNotificationService {
         object ClientNotAuthorized: ConnectionError()  // 39507
         data class Failed(val errorCode: Int?): ConnectionError()
     }
-}
 
-fun ExposureInformation.toExposure(): Exposure {
-    Timber.d(this.toString())
+    // Specifies what was canceled when enable result is Canceled (these are only used for Huawei Contact Shield)
+    sealed class EnableStep {
+        object UserConsent: EnableStep()
+        object LocationPermission: EnableStep()
+    }
 
-    return Exposure(
-        detectedDate = ZonedDateTime.ofInstant(Instant.ofEpochMilli(this.dateMillisSinceEpoch), ZoneOffset.UTC),
-        totalRiskScore = this.totalRiskScore,
-        createdDate = ZonedDateTime.now()
-    )
+    // Provides methods to check and resolve exposure system availability in the UI
+    interface AvailabilityResolver {
+        fun isSystemAvailable(context: Context): Int
+        fun isUserResolvableError(errorCode: Int): Boolean
+        fun showErrorDialogFragment(activity: Activity, errorCode: Int, requestCode: Int, cancelListener: (dialog: DialogInterface) -> Unit): Boolean
+    }
 }