+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Manual exposure check allowed when checks have been delayed 24h
+- Display exposure notification count on home screen and notification list on exposure detail screen
+- Municipality data that controls the availability of Omaolo contact request
+
+### Changed
+- App structure and build properties to support Huawei Contact Shield build flavor
+- Dependency updates
 
 ## [1.2.1](https://github.com/THLfi/koronavilkku-android/compare/v1.2.0...v1.2.1) - 2020-10-29